requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiokafka==0.10.0
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.0
click==8.1.7
colorlog==6.8.2
constantly==23.10.4
croniter==2.0.1
cryptography==42.0.1
daphne==3.0.2
Django==3.2
djangorestframework==3.14.0
dnspython==2.5.0
eventlet==0.35.0
faust-aioeventlet==0.6
faust-streaming==0.10.22
frozenlist==1.4.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
intervaltree==3.1.0
mode-streaming==0.3.5
multidict==6.0.4
mypy-extensions==1.0.0
opentracing==2.4.0
packaging==23.2
psycopg2-binary==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
service-identity==24.1.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
terminaltables==3.1.10
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzlocal==5.2
venusian==3.1.0
yarl==1.9.4
zope.interface==6.1
